refactor: unique_sorted_additional_folders を path_utils へ移設し相互依存を解消 (#87)
scan → fingerprint（トークン生成）と fingerprint → scan（フォルダ正規化）の
相互 import を、共有ヘルパーの下層（path_utils）移設で一方向化。
残る fingerprint → scan 依存は #[cfg(test)] の build_fingerprint のみ。
