+++ src-tauri/src/commands/ghost/path_utils.rs
@@ -1,5 +1,22 @@
-use std::path::Path;
+use std::path::{Path, PathBuf};
 
 pub(crate) fn normalize_path(path: &Path) -> String {
     path.to_string_lossy().replace('\\', "/").to_lowercase()
+}
+
+pub(crate) fn unique_sorted_additional_folders(
+    additional_folders: &[String],
+) -> Vec<(String, PathBuf, String)> {
+    let mut folders = additional_folders
+        .iter()
+        .map(|folder| {
+            let path = PathBuf::from(folder);
+            let normalized = normalize_path(&path);
+            (folder.clone(), path, normalized)
+        })
+        .collect::<Vec<_>>();
+
+    folders.sort_by(|a, b| a.2.cmp(&b.2));
+    folders.dedup_by(|a, b| a.2 == b.2);
+    folders
 }
